Project/deepreason/requirements.txt
flask
gunicorn
python-dotenv
werkzeug
pdfplumber
openai